refactor: extract throttle utilities to shared package
Eliminate client/server code duplication:
- Create packages/core/src/throttle-utils.ts
- Export SupportedWindow type + computeWindowStartMs + constants
- Single source of truth for window duration calculations

Server (proxy) and client (dashboard) both import from @better-ccflare/core.
Removes ~99 lines of duplicated code.
Prevents formula drift between server enforcement and client display.

Monthly window calendar math (28/29/30/31 days) now centralized.
Type-safe shared utilities with automatic sync.

Addresses Greptile P2 maintainability concern.
